Exception handling - deelete

+++ src/main/java/com/meuportfolio/backend/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.meuportfolio.backend.entities.User;
 import com.meuportfolio.backend.repositories.UserRepository;
+import com.meuportfolio.backend.services.exceptions.DatabaseException;
 import com.meuportfolio.backend.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -15,24 +18,40 @@ public class UserService {
     @Autowired
     private UserRepository repository;
 
-    public List<User>findAll(){
+    public List<User> findAll() {
         return repository.findAll();
     }
-    public User findById(Long id){
-       Optional<User>obj = repository.findById(id);
-       return obj.orElseThrow(() -> new ResourceNotFoundException(id));
+
+    public User findById(Long id) {
+        Optional<User> obj = repository.findById(id);
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
-    public User insert(User obj){
+
+    public User insert(User obj) {
         return repository.save(obj);
     }
-    public void delete(Long id){
-        repository.deleteById(id);
+
+    public void delete(Long id) {
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+
+        } catch (EmptyResultDataAccessException e) {
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
-    public User update(Long id, User obj){
+
+    public User update(Long id, User obj) {
         User entity = repository.getReferenceById(id);
         updateData(entity, obj);
         return repository.save(entity);
     }
+
     private void updateData(User entity, User obj) {
         entity.setName(obj.getName());
         entity.setEmail(obj.getEmail());
